+++ HelloWorldGWT/src/br/unicamp/mc437/correio/spec/prov/ICorreio.java
@@ -3,7 +3,10 @@ package br.unicamp.mc437.correio.spec.prov;
 
 public interface ICorreio{
 
-	public float valorFrete (  ); 
-	public String obterProtocolo ( int id_compra ); 
+	public double valorFrete (  ); 
+	
+	public String obterProtocolo ( int id_compra );
+	
 	public String acompanharStatus ( String protocolo ); 
+	
 }
